# openapi/cafe-kiosk-api-v1.yaml
...
openapi: 3.0.3
info:
  title: CafeKiosk API
  version: "1.0.0"
  description: TDD로 개발하는 카페 키오스크 API 서버
servers:
  - url: /
    description: Local server
paths:
...
  /api/v1/orders:
    post:
      tags:
        - Order
      summary: 신규 주문 등록
      description: 상품 번호 목록을 받아 신규 주문을 등록합니다.
      operationId: createOrder
      requestBody:
        description: 신규 주문에 포함될 상품 번호 목록
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
      responses:
        '201':
          description: 주문 등록 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/orders/{orderId}/status:
    patch:
      tags:
        - Order
      summary: 주문 상태 변경
      description: 특정 주문의 상태를 변경합니다.
      operationId: updateOrderStatus
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: 주문 ID
      requestBody:
        description: 변경할 주문 상태 정보
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderStatusUpdateRequest'
      responses:
        '200':
          description: 주문 상태 변경 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 주문을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ProductResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 상품 ID
        productNumber:
          type: string
          description: 상품 번호
        name:
          type: string
          description: 상품명
        price:
          type: integer
          format: int32
          description: 가격
        sellingStatus:
          $ref: '#/components/schemas/SellingStatus'
    SellingStatus:
      type: string
      enum:
        - SELLING
        - HOLD
        - STOP_SELLING
      description: |
        상품 판매 상태:
        * `SELLING`: 판매중
        * `HOLD`: 판매 보류
        * `STOP_SELLING`: 판매 중지
    OrderCreateRequest:
      type: object
      properties:
        productNumbers:
          type: array
          items:
            type: string
          description: 주문할 상품의 상품 번호 목록
      required:
        - productNumbers
    OrderResponse:
      type: object
      x-lombok-builder: true # 이 부분을 추가합니다.
      properties:
        id:
          type: integer
          format: int64
          description: 주문 ID
        totalPrice:
          type: integer
          format: int32
          description: 총 주문 금액
        registeredDateTime:
          type: string
          format: date-time
          description: 주문 등록 일시
        orderStatus: # 이 부분을 추가합니다.
          $ref: '#/components/schemas/OrderStatus' # OrderStatus 스키마를 참조합니다.
        products:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
    OrderStatus: # 이 부분을 추가합니다.
      type: string
      enum:
        - INIT
        - CANCELED
        - COMPLETED
      description: |
        주문 상태:
        * `INIT`: 주문 접수
        * `CANCELED`: 주문 취소
        * `COMPLETED`: 처리 완료
    OrderStatusUpdateRequest:
      type: object
      properties:
        orderStatus:
          $ref: '#/components/schemas/OrderStatus'
      required:
        - orderStatus
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: 에러 코드
        message:
          type: string
          description: 에러 메시지
...